refactor: debounce定数の共通化・テスト修正・performance静的データ分離
- PERSIST_DEBOUNCE_MS を src/constants/persist.ts に集約（7箇所の重複定義を解消）
- settings.test.ts の JSON.parse を JSON5.parse に修正（settings.json5 移行対応）
- deck.test.ts から削除済み loadWallpaper() の呼び出しを除去
- 未使用の src/defaults/appearances.json5 を削除
- performance.ts の静的データ（FIELD_META, プリセット, 補間関数等 ~590行）を
  performanceData.ts に分離し、store ロジックを見通しやすく整理
