+++ src/constants/persist.ts
@@ -1,0 +1,2 @@
+/** Debounce delay (ms) for all file/localStorage persist operations. */
+export const PERSIST_DEBOUNCE_MS = 300
