--- src/defaults/appearances.json5
@@ -1,3 +1,0 @@
-// NoteDeck settings.json5 デフォルト
-// スカラー preferences / アプリ状態
-{}
